created types for billing and renamed billingUtil to billing/index
