+++ src/services/billingService.ts
@@ -1,42 +1,14 @@
 import { getEndpoint, getPaymentsUrl } from 'utils/common/apiUtil';
 import { getToken } from 'utils/common/key';
 import { setData, LS_KEYS } from 'utils/storage/localStorage';
-import { convertToHumanReadable } from 'utils/billingUtil';
+import { convertToHumanReadable } from 'utils/billing';
 import HTTPService from './HTTPService';
 import { logError } from 'utils/sentry';
 import { getPaymentToken } from './userService';
+import { Plan, PaymentActionType, Subscription } from 'types/billing';
 
 const ENDPOINT = getEndpoint();
 
-enum PaymentActionType {
-    Buy = 'buy',
-    Update = 'update',
-}
-export interface Subscription {
-    id: number;
-    userID: number;
-    productID: string;
-    storage: number;
-    originalTransactionID: string;
-    expiryTime: number;
-    paymentProvider: string;
-    attributes: {
-        isCancelled: boolean;
-    };
-    price: string;
-    period: string;
-}
-export interface Plan {
-    id: string;
-    androidID: string;
-    iosID: string;
-    storage: number;
-    price: string;
-    period: string;
-    stripeID: string;
-}
-
-export const FREE_PLAN = 'free';
 class billingService {
     public async getPlans(): Promise<Plan[]> {
         try {
