+++ src/types/billing/index.ts
@@ -1,0 +1,29 @@
+export enum PaymentActionType {
+    Buy = 'buy',
+    Update = 'update',
+}
+export const FREE_PLAN = 'free';
+
+export interface Subscription {
+    id: number;
+    userID: number;
+    productID: string;
+    storage: number;
+    originalTransactionID: string;
+    expiryTime: number;
+    paymentProvider: string;
+    attributes: {
+        isCancelled: boolean;
+    };
+    price: string;
+    period: string;
+}
+export interface Plan {
+    id: string;
+    androidID: string;
+    iosID: string;
+    storage: number;
+    price: string;
+    period: string;
+    stripeID: string;
+}
